Implement summoners name updater
